Run libuca-related tests in container where libuca is installed
