+++ concert/tests/integration/scenarios/test_367.py
@@ -5,11 +5,8 @@ from concert.quantities import q
 class TestIssue367(TestCase):
 
     async def test_degree_conversion(self):
-        try:
-            from concert.devices.cameras.uca import Camera
-            camera = await Camera("mock")
-        except Exception as err:
-            self.skipTest(str(err))
+        from concert.devices.cameras.uca import Camera
+        camera = await Camera(name="mock")
 
         await camera.set_degree_value(q.Quantity(5, q.celsius))
         self.assertEqual((await camera.get_degree_value()).magnitude, 5.0)
