--- concert/tests/regressions/test_400.py
@@ -1,17 +1,0 @@
-from concert.tests import TestCase
-
-
-class TestIssue400(TestCase):
-
-    async def test_multiple_cameras(self):
-        try:
-            from concert.devices.cameras.uca import Camera
-
-            cam_a = await Camera(name='mock')
-            cam_b = await Camera(name='file')
-            cam_c = await Camera(name='mock')
-        except Exception as err:
-            self.skipTest(str(err))
-
-        self.assertEqual("degree_value" in dir(cam_b), False)
-        self.assertEqual("path" in dir(cam_c), False)
